678: Add contracted hours to workload points calculation
Remove working_hours table, and include contracted hours in the workload_owner
table. Update unit and integration tests to account for working_hours
table no longer existing.

Update contracted_hours and reductions to be float, as opposed to int

+++ migrations/app/20170620142722_app_reductions.js
@@ -3,7 +3,7 @@ exports.up = function (knex, Promise) {
   return knex.schema.createTable('reductions', function (table) {
     table.increments('id')
     table.integer('workload_owner_id').unsigned().notNullable().references('workload_owner.id')
-    table.integer('hours').unsigned().notNullable()
+    table.float('hours').unsigned().notNullable()
     table.timestamp('effective_from').notNullable()
     table.timestamp('effective_to').notNullable()
   }).catch(function (error) {
